services/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose==3.3.0
python-multipart==0.0.9
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-jaeger==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-httpx==0.46b0
